footer transfer to navbar
