+++ client/src/App.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import './App.css';
 import Header from './components/header/Header';
-import Footer from './components/footer/Footer';
+import Footer from './components/navbar/Navbar';
 import Main from './components/main/Main';
 // import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
